Upgrade GCP client, add initer/closer interfaces.
Upgrade our GCP KMV client to the latest release. This requires us to be
able to close the client when we're done with it.

Add Initer and Closer interfaces to support this. Clients that don't
need them don't need to implement them, and we can use interface checks
to safely create/call them.